Zapisywanie z pomijaniem
